requirements.txt
flask
google-generativeai
langchain
faiss-cpu
pypdf
python-dotenv


